pyproject.toml
[project]
name = "onvoke-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx2pdf>=0.1.8",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "huggingface-hub>=0.33.1",
    "langchain>=0.3.26",
    "langchain-core>=0.3.69",
    "langchain-google-genai>=2.0.10",
    "langchain-text-splitters>=0.3.8",
    "langgraph>=0.5.3",
    "langgraph-prebuilt>=0.5.2",
    "langgraph-sdk>=0.1.73",
    "langsmith>=0.4.7",
    "google-generativeai>=0.8.0",
    "langchain-google-genai>=2.0.0",
    "llama-cloud-services>=0.6.50",
    "llama-index>=0.12.49",
    "llama-index-vector-stores-supabase>=0.3.0",
    "llama-parse>=0.6.50",
    "markdown>=3.8.2",
    "markdown-it-py>=3.0.0",
    "markdown-pdf>=1.7",
    "numpy>=2.3.1",
    "pandas>=2.3.1",
    "pdflatex>=0.1.3",
    "pillow>=11.3.0",
    "pydantic>=2.11.7",
    "pydantic-core>=2.33.2",
    "pymupdf>=1.25.3",
    "pypandoc>=1.15",
    "pypdf>=5.8.0",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "python-multipart>=0.0.20",
    "loguru>=0.7.0",
    "realtime>=2.6.0",
    "regex>=2024.11.6",
    "reportlab>=4.4.2",
    "requests>=2.32.4",
    "scikit-learn>=1.7.1",
    "supabase>=2.17.0",
    "torch>=2.7.1",
    "uvicorn>=0.35.0",
    "weasyprint>=65.1",
    "wkhtmltopdf>=0.2",
    "yarl>=1.20.1",
    "zopfli>=0.2.3.post1",
    "zstandard>=0.23.0",
]
